+++ packages/bpk-component-datepicker/src/BpkDatepicker.js
@@ -5,6 +5,18 @@ import React, { PropTypes, Component } from 'react';
 
 import './bpk-datepicker.scss';
 
+const KEYCODES = {
+  ENTER: 13,
+  SPACEBAR: 32,
+};
+
+const onKeyEvent = (keyCode, callback) => (e) => {
+  if (e.keyCode === keyCode) {
+    e.preventDefault();
+    callback();
+  }
+};
+
 class BpkDatepicker extends Component {
   constructor(props) {
     super(props);
@@ -78,50 +90,48 @@ class BpkDatepicker extends Component {
     delete rest.onDateSelect;
 
     return (
-      <div>
-        <span
-          className="bpk-datepicker__input-description"
-          aria-live="polite"
-        >{ formatDateFull(this.state.date) }</span>
-        <BpkPopover
-          target={
-            <BpkInput
-              id={id}
-              name={`${id}_input`}
-              value={formatDate(this.state.date)}
-              onClick={this.onOpen}
-              onFocus={this.onOpen}
-              className="bpk-datepicker__input"
-              {...inputProps}
-            />
-          }
-          onClose={this.onClose}
-          isOpen={this.state.isOpen}
-          closeButtonText="Close"
-          title={popoverLabel}
-          closeOnOutsideClickExceptTarget
-          {...rest}
-        >
-          <BpkCalendar
-            changeMonthLabel={changeMonthLabel}
-            DateComponent={DateComponent}
-            dateModifiers={dateModifiers}
-            daysOfWeek={daysOfWeek}
-            enableSelection
-            formatDateFull={formatDateFull}
-            formatMonth={formatMonth}
-            id={`${id}_calendar`}
-            initialSelectedDate={this.state.date}
-            markOutsideDays={markOutsideDays}
-            markToday={markToday}
-            maxDate={maxDate}
-            minDate={minDate}
-            onDateSelect={this.onDateSelect}
-            showWeekendSeparator={showWeekendSeparator}
-            weekStartsOn={weekStartsOn}
+      <BpkPopover
+        target={
+          <BpkInput
+            id={id}
+            name={`${id}_input`}
+            value={formatDate(this.state.date)}
+            onClick={this.onOpen}
+            onFocus={this.onOpen}
+            onKeyDown={onKeyEvent(KEYCODES.ENTER, this.onOpen)}
+            onKeyUp={onKeyEvent(KEYCODES.SPACEBAR, this.onOpen)}
+            className="bpk-datepicker__input"
+            aria-live="assertive"
+            aria-atomic="true"
+            aria-label={formatDateFull(this.state.date)}
+            {...inputProps}
           />
-        </BpkPopover>
-      </div>
+        }
+        onClose={this.onClose}
+        isOpen={this.state.isOpen}
+        closeButtonText="Close"
+        title={popoverLabel}
+        {...rest}
+      >
+        <BpkCalendar
+          changeMonthLabel={changeMonthLabel}
+          DateComponent={DateComponent}
+          dateModifiers={dateModifiers}
+          daysOfWeek={daysOfWeek}
+          enableSelection
+          formatDateFull={formatDateFull}
+          formatMonth={formatMonth}
+          id={`${id}_calendar`}
+          initialSelectedDate={this.state.date}
+          markOutsideDays={markOutsideDays}
+          markToday={markToday}
+          maxDate={maxDate}
+          minDate={minDate}
+          onDateSelect={this.onDateSelect}
+          showWeekendSeparator={showWeekendSeparator}
+          weekStartsOn={weekStartsOn}
+        />
+      </BpkPopover>
     );
   }
 }